requirements.txt
amqp==5.2.0
asgiref==3.8.1
astroid==2.15.8
attrs==24.2.0
Authlib==1.3.2
autopep8==2.3.1
billiard==4.2.0
black==24.8.0
build==1.2.1
celery==5.4.0
celery-progress==0.4
certifi==2024.7.4
cffi==1.17.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==43.0.1
distlib==0.3.8
dj-rest-auth==6.0.0
dj-stripe==2.8.4
Django==5.1.1
django-allauth==64.1.0
django-anymail==11.1
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-filter==24.3
django-health-check==3.18.3
django-hijack==3.6.0
django-otp==1.5.2
django-redis==5.4.0
django-timezone-field==7.0
django-waffle==4.1.0
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
docutils==0.21.2
drf-spectacular==0.27.2
fido2==1.1.3
filelock==3.15.4
flake8==7.1.1
gunicorn==23.0.0
identify==2.6.0
idna==3.7
inflection==0.5.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
lazy-object-proxy==1.10.0
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
pip-tools==7.4.1
platformdirs==4.2.2
pre-commit==3.8.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
pyotp==2.9.0
pypng==0.20220715.0
pyproject_hooks==1.1.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
qrcode==7.4.2
redis==5.0.8
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
ruff==0.6.2
setuptools==74.0.0
six==1.16.0
sqlparse==0.5.1
stripe==4.2.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
wheel==0.44.0
wrapt==1.16.0
